refactor(monorepo): make the project a monorepo and setup config
