+++ .gitignore
@@ -1,3 +1,4 @@
 .history
 build
 node_modules
+.idea
