+++ lerna.json
@@ -1,0 +1,12 @@
+{
+  "packages": ["packages/*"],
+  "ignoreChanges": ["packages/*/test/**", "**/*.md"],
+  "version": "independent",
+  "npmClient": "yarn",
+  "useWorkspaces": true,
+  "command": {
+    "publish": {
+      "conventionalCommits": true
+    }
+  }
+}
