+++ packages/button/README.md
@@ -1,0 +1,11 @@
+# `button`
+
+> TODO: description
+
+## Usage
+
+```
+const button = require('button');
+
+// TODO: DEMONSTRATE API
+```
