+++ packages/button/bw-button.ts
@@ -1,0 +1,3 @@
+import { BwButton} from './src/BwButton';
+
+customElements.define('bw-button', BwButton);
