+++ packages/button/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "button",
+  "version": "0.0.1",
+  "description": "A bulma.io based web-component button",
+  "keywords": [
+    "web-components",
+    "button",
+    "bulma-web",
+    "bulma"
+  ],
+  "author": "Ramon Osenga <ramon.osenga@gmail.com>",
+  "homepage": "https://github.com/Ramon92/bulma-web#readme",
+  "license": "MIT",
+  "main": "index.js",
+  "directories": {
+    "lib": "lib",
+    "test": "__tests__"
+  },
+  "publishConfig": {
+    "registry": "https://registry.yarnpkg.com"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Ramon92/bulma-web.git"
+  },
+  "scripts": {
+    "test": "echo \"Error: run tests from root\" && exit 1"
+  },
+  "bugs": {
+    "url": "https://github.com/Ramon92/bulma-web/issues"
+  },
+  "dependencies": {
+    "@lion/button": "^0.3.11",
+    "bulma": "^0.7.5",
+    "lit-element": "^2.2.1"
+  }
+}
