--- src/components/button/bw-button.ts
@@ -1,3 +1,0 @@
-import { BwButton} from './BwButton';
-
-customElements.define('bw-button', BwButton);
